Move checker to separate translation unit
